requirements.txt
annotated-types==0.7.0
dnspython==2.7.0
email_validator==2.2.0
greenlet==3.1.1
idna==3.10
mysql-connector-python==9.2.0
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL==1.1.1
SQLAlchemy==2.0.37
sqlmodel==0.0.22
typing_extensions==4.12.2
